Reorganising neutron_stars module
